Avoid multiple writes and prints on distributed settings
